requirements.txt
# ============================================
# TrouveUnCadeau.xyz - Requirements
# Architecture Multi-LLM Optimisée
# Compatible Render.com
# ============================================

# === Web Framework ===
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.12

# === LLM Providers ===
# Together.ai - Provider principal (90% trafic)
langchain-together==0.1.3

# Google Gemini - Provider secondaire (8% trafic)
langchain-google-genai==1.0.10
google-generativeai==0.8.3

# Anthropic Claude - Provider tertiaire (2% trafic)
langchain-anthropic==0.2.3
anthropic==0.39.0

# OpenAI - Embeddings uniquement
langchain-openai==0.2.2
openai==1.54.3

# === LangChain Core ===
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5

# === Vector Store ===
faiss-cpu==1.8.0.post1
numpy==1.26.4

# === Cache ===
redis==5.1.1
hiredis==3.0.0  # Accélère Redis

# === Data Processing ===
pandas==2.2.3
pyairtable==2.3.3

# === Monitoring & Logging ===
python-json-logger==2.0.7
structlog==24.4.0

# === Utils ===
requests==2.32.3
httpx==0.27.2
tenacity==9.0.0  # Retry logic

# === Production ===
gunicorn==23.0.0  # Pour Render.com
